Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-in-setTimeout-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected end of script

PASS window.onerror - compile error in setTimeout 
PASS window.onerror - compile error in setTimeout (column) 
